+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Edit Profile view navigation bar area. [#329](https://github.com/verse-pbc/issues/issues/329)
 - Added confirmation prompt when leaving a group. [#331](https://github.com/verse-pbc/issues/issues/331)
 - Added prominent create community button to the group list view.
+- Added a welcome screen and updated the old login screen.
 
 ### Internal Changes
 - Standardized on FVM for Flutter version management, removing mise configuration.